+++ docs/changes.rst
@@ -36,6 +36,7 @@ Unreleased
 - New: ``sheldon``
 - New: ``thegutters``
 - Update ``ctrlaltdelete`` to work after source site changes.
+- Update ``dieselsweetiesweb`` to work after source site changes.
 - Update ``eon`` with new source site.
 - Update ``lunch`` with new source site.
 - Update ``sometingpositive`` to get all releases.